+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/MasterDataValidator.cs
@@ -29,6 +29,7 @@ using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.OtherProduction
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.OwnProduction;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Production;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.SupplyToGrid;
+using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.SurplusProduction;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.TotalConsumption;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.VEProduction;
 using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.WholesaleServices;
